# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Irish Payroll Management & Compliance System",
  "description": "Portfolio prototype. Statutory values are versioned rules with official sources; unverified rules are blocked from calculation. Not certified payroll software.",
  "version": "1.0.0"
 },
 "paths": {
...
  "/payroll-runs": {
   "post": {
    "tags": [
     "payroll"
    ],
    "summary": "Create Run",
    "operationId": "create_run_payroll_runs_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RunIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RunOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "payroll"
    ],
    "summary": "List Runs",
    "operationId": "list_runs_payroll_runs_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "company_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Company Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/RunOut"
         },
         "title": "Response List Runs Payroll Runs Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/payroll-runs/{run_id}": {
   "get": {
    "tags": [
     "payroll"
    ],
    "summary": "Get Run",
    "operationId": "get_run_payroll_runs__run_id__get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Run Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RunOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/payroll-runs/{run_id}/inputs": {
   "post": {
    "tags": [
     "payroll"
    ],
    "summary": "Add Inputs",
    "operationId": "add_inputs_payroll_runs__run_id__inputs_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Run Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/InputsIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/payroll-runs/{run_id}/calculate": {
   "post": {
    "tags": [
     "payroll"
    ],
    "summary": "Calculate",
    "operationId": "calculate_payroll_runs__run_id__calculate_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Run Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/payroll-runs/{run_id}/validate": {
   "post": {
    "tags": [
     "payroll"
    ],
    "summary": "Validate",
    "operationId": "validate_payroll_runs__run_id__validate_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Run Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/payroll-runs/{run_id}/approve": {
   "post": {
    "tags": [
     "payroll"
    ],
    "summary": "Approve",
    "operationId": "approve_payroll_runs__run_id__approve_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Run Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/payroll-runs/{run_id}/reverse": {
   "post": {
    "tags": [
     "payroll"
    ],
    "summary": "Reverse",
    "operationId": "reverse_payroll_runs__run_id__reverse_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Run Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ReasonIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RunOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/payroll-runs/{run_id}/results": {
   "get": {
    "tags": [
     "payroll"
    ],
    "summary": "Results",
    "operationId": "results_payroll_runs__run_id__results_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Run Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/payroll-results/{result_id}/explain": {
   "get": {
    "tags": [
     "payroll"
    ],
    "summary": "Explain",
    "description": "'Why was this employee's PAYE \u20acx?' - full structured calculation trace.",
    "operationId": "explain_payroll_results__result_id__explain_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "result_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Result Id"
      }
     },
     {
      "name": "component",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Component"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/payroll-runs/{run_id}/reconciliation": {
   "get": {
    "tags": [
     "payroll"
    ],
    "summary": "Reconciliation",
    "operationId": "reconciliation_payroll_runs__run_id__reconciliation_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Run Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/payroll-runs/{run_id}/journal": {
   "get": {
    "tags": [
     "payroll"
    ],
    "summary": "Journal",
    "operationId": "journal_payroll_runs__run_id__journal_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Run Id"
      }
     },
     {
      "name": "fmt",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "json",
       "title": "Fmt"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/payroll-runs/{run_id}/export.xlsx": {
   "get": {
    "tags": [
     "payroll"
    ],
    "summary": "Export Xlsx",
    "operationId": "export_xlsx_payroll_runs__run_id__export_xlsx_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Run Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/employees/{employee_id}/payslip": {
   "get": {
    "tags": [
     "payroll"
    ],
    "summary": "Payslip",
    "operationId": "payslip_employees__employee_id__payslip_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "employee_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Employee Id"
      }
     },
     {
      "name": "run_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Run Id"
      }
     },
     {
      "name": "fmt",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "html",
       "title": "Fmt"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/exceptions": {
   "get": {
    "tags": [
     "payroll"
    ],
    "summary": "List Exceptions",
    "operationId": "list_exceptions_exceptions_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "run_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Run Id"
      }
     },
     {
      "name": "severity",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Severity"
      }
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/exceptions/{exception_id}/resolve": {
   "post": {
    "tags": [
     "payroll"
    ],
    "summary": "Resolve",
    "operationId": "resolve_exceptions__exception_id__resolve_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "exception_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Exception Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ExceptionResolveIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "DeductionIn": {
    "properties": {
     "code": {
      "type": "string",
      "title": "Code"
     },
     "amount": {
      "anyOf": [
       {
        "type": "number",
        "minimum": 0.0
       },
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       }
      ],
      "title": "Amount"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     }
    },
    "type": "object",
    "required": [
     "code",
     "amount"
    ],
    "title": "DeductionIn"
   },
   "EarningIn": {
    "properties": {
     "type": {
      "type": "string",
      "title": "Type"
     },
     "amount": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       }
      ],
      "title": "Amount"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "hours": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Hours"
     }
    },
    "type": "object",
    "required": [
     "type",
     "amount"
    ],
    "title": "EarningIn"
   },
...
   "ExceptionResolveIn": {
    "properties": {
     "note": {
      "type": "string",
      "minLength": 3,
      "title": "Note"
     },
     "status": {
      "type": "string",
      "enum": [
       "RESOLVED",
       "ACKNOWLEDGED"
      ],
      "title": "Status",
      "default": "RESOLVED"
     }
    },
    "type": "object",
    "required": [
     "note"
    ],
    "title": "ExceptionResolveIn"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "InputsIn": {
    "properties": {
     "employee_id": {
      "type": "integer",
      "title": "Employee Id"
     },
     "earnings": {
      "items": {
       "$ref": "#/components/schemas/EarningIn"
      },
      "type": "array",
      "title": "Earnings",
      "default": []
     },
     "deductions": {
      "items": {
       "$ref": "#/components/schemas/DeductionIn"
      },
      "type": "array",
      "title": "Deductions",
      "default": []
     }
    },
    "type": "object",
    "required": [
     "employee_id"
    ],
    "title": "InputsIn"
   },
   "ReasonIn": {
    "properties": {
     "reason": {
      "type": "string",
      "minLength": 5,
      "title": "Reason"
     }
    },
    "type": "object",
    "required": [
     "reason"
    ],
    "title": "ReasonIn"
   },
   "RunIn": {
    "properties": {
     "company_id": {
      "type": "integer",
      "title": "Company Id"
     },
     "tax_year": {
      "type": "integer",
      "title": "Tax Year"
     },
     "frequency": {
      "type": "string",
      "enum": [
       "WEEKLY",
       "FORTNIGHTLY",
       "MONTHLY"
      ],
      "title": "Frequency"
     },
     "payroll_period": {
      "type": "integer",
      "maximum": 53.0,
      "minimum": 1.0,
      "title": "Payroll Period"
     },
     "period_start": {
      "type": "string",
      "format": "date",
      "title": "Period Start"
     },
     "period_end": {
      "type": "string",
      "format": "date",
      "title": "Period End"
     },
     "payment_date": {
      "type": "string",
      "format": "date",
      "title": "Payment Date"
     }
    },
    "type": "object",
    "required": [
     "company_id",
     "tax_year",
     "frequency",
     "payroll_period",
     "period_start",
     "period_end",
     "payment_date"
    ],
    "title": "RunIn"
   },
   "RunOut": {
    "properties": {
     "payroll_run_id": {
      "type": "integer",
      "title": "Payroll Run Id"
     },
     "company_id": {
      "type": "integer",
      "title": "Company Id"
     },
     "tax_year": {
      "type": "integer",
      "title": "Tax Year"
     },
     "payroll_period": {
      "type": "integer",
      "title": "Payroll Period"
     },
     "frequency": {
      "type": "string",
      "title": "Frequency"
     },
     "period_start": {
      "type": "string",
      "format": "date",
      "title": "Period Start"
     },
     "period_end": {
      "type": "string",
      "format": "date",
      "title": "Period End"
     },
     "payment_date": {
      "type": "string",
      "format": "date",
      "title": "Payment Date"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "run_type": {
      "type": "string",
      "title": "Run Type"
     },
     "totals": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Totals"
     },
     "calc_seconds": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Calc Seconds"
     }
    },
    "type": "object",
    "required": [
     "payroll_run_id",
     "company_id",
     "tax_year",
     "payroll_period",
     "frequency",
     "period_start",
     "period_end",
     "payment_date",
     "status",
     "run_type",
     "totals",
     "calc_seconds"
    ],
    "title": "RunOut"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "type": "oauth2",
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "/auth/token"
     }
    }
   }
  }
 }
}